Move parseAnnotatedPncUpdateDatasetXml function into root lib folder
