+++ packages/core/phase3/tests/helpers/mockUpdatesInPnc.ts
@@ -1,9 +1,9 @@
 import { isError } from "@moj-bichard7/common/types/Result"
 
+import parseAnnotatedPncUpdateDatasetXml from "../../../lib/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
 import getMessageType from "../../../phase1/lib/getMessageType"
 import addMockToPnc from "../../../phase1/tests/helpers/addMockToPnc"
 import clearMocksInPnc from "../../../phase1/tests/helpers/clearMocksInPnc"
-import parseAnnotatedPncUpdateDatasetXml from "../../../phase2/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
 import { parsePncUpdateDataSetXml } from "../../../phase2/parse/parsePncUpdateDataSetXml"
 import getPncErrorMessages from "./getPncErrorMessages"
 import getPncOperationsFromPncUpdateDataset from "./getPncOperationsFromPncUpdateDataset"
