+++ packages/ui/src/utils/parseHearingOutcome.test.ts
@@ -1,11 +1,11 @@
 import parseAhoXml from "@moj-bichard7/core/lib/parse/parseAhoXml/parseAhoXml"
-import parseAnnotatedPncUpdateDatasetXml from "@moj-bichard7/core/phase2/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
+import parseAnnotatedPncUpdateDatasetXml from "@moj-bichard7/core/lib/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
 import fs from "fs"
 import dummyAho from "../../test/test-data/AnnotatedHO1.json"
 import parseHearingOutcome from "./parseHearingOutcome"
 
 jest.mock("@moj-bichard7/core/lib/parse/parseAhoXml/parseAhoXml")
-jest.mock("@moj-bichard7/core/phase2/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml")
+jest.mock("@moj-bichard7/core/lib/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml")
 
 beforeEach(() => {
   ;(parseAhoXml as jest.Mock).mockImplementation(
@@ -13,7 +13,7 @@ beforeEach(() => {
   )
   ;(parseAnnotatedPncUpdateDatasetXml as jest.Mock).mockImplementation(
     jest.requireActual(
-      "@moj-bichard7/core/phase2/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
+      "@moj-bichard7/core/lib/parse/parseAnnotatedPncUpdateDatasetXml/parseAnnotatedPncUpdateDatasetXml"
     ).default
   )
 })
